Preliminary work on separating switch specific implementation and common code Add initial support for cisco 9300 switch Only necessary commands used for test results implemented at this point, additional reporting only commands not present
